backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
pydantic==2.10.4
pydantic-settings==2.7.0
paramiko==3.5.0
PyYAML==6.0.2
python-pam==2.0.2
PyJWT==2.10.1
python-multipart==0.0.20
hypothesis==6.122.3
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
